+++ lint/rules/index.js
@@ -1,0 +1,4 @@
+require('./metadata.rules');
+require('./schema.rules');
+require('./naming.rules');
+require('./dependency.rules');